feature: 퀴즈 repository 코드 수정
