+++ src/main/java/com/goldbalance/dive/domain/article/repository/article/ArticleCustomRepositoryImpl.java
@@ -4,7 +4,6 @@ import static com.goldbalance.dive.domain.article.domain.QArticle.*;
 import static com.goldbalance.dive.domain.article.domain.QQuiz.*;
 
 import com.goldbalance.dive.domain.article.domain.Article;
-import com.goldbalance.dive.domain.article.domain.Quiz;
 import com.goldbalance.dive.domain.article.dto.request.ArticleQueryOption;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
@@ -21,14 +20,6 @@ public class ArticleCustomRepositoryImpl implements ArticleCustomRepository {
         return queryFactory
                 .selectFrom(article)
                 .where(containsKeyword(queryOption))
-                .fetch();
-    }
-
-    @Override
-    public List<Quiz> searchQuiz(Long articleId) {
-        return queryFactory
-                .selectFrom(quiz)
-                .where(quiz.article.id.eq(articleId))
                 .fetch();
     }
