+++ src/main/java/com/goldbalance/dive/domain/article/repository/quiz/QuizRepository.java
@@ -3,4 +3,4 @@ package com.goldbalance.dive.domain.article.repository.quiz;
 import com.goldbalance.dive.domain.article.domain.Quiz;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuizRepository extends JpaRepository<Quiz, Long> {}
+public interface QuizRepository extends JpaRepository<Quiz, Long>, QuizCustomRepository {}
